Ensure that deployment host cache is owned appropriately
When pulling the packaged venv to the deployment host, the creation
of the local file on the deploy host is done using the user that is
running ansible. If ansible is being run by a non-root user, then
the folder that's created to store the files will not have the right
permissions and the fetch will fail.

As such, when acting on the deploy host we should always ensure that
we provide the correct rights to the user running ansible. We do this
by using a lookup to figure out which user is executing the playbook,
then setting the ownership of the folder to that user. We also use a
lookup to determine that user's home directory and default to using
a subdirectory of that folder for the cache. Both lookups have options
to fall back to in case the environment variables used are not available.