Ensure venv is extracted if the target folder was not present
In the situation where a deployer may have removed the venv folder
on the target host, but has not removed the the tgz/checksum file,
the venv will not be extracted again.

This patch ensures that if the venv folder was not there and gets
created by the task then it will extract the tgz into the folder.