Ensure python/paths are updated if folder was not present
In the situation where a deployer may have removed the venv folder
on the target host, but has not removed the the tgz/checksum file,
the venv will get extracted, but the python/paths will not get
updated.

This patch ensures that if the venv folder was not there and gets
created by the task then it will also update the python/paths.

Change-Id: I8858e16a8c128e7d7f44c5e46dc8520aa5bc835b