Revert the use of wheel synchronizing
When using this role, a playbook can be used to orchestrate
doing all the builds on a single host. There is no need to
orchestrate this inside the role.

In order to gain the efficiencies of building everything on
a single host, a playbook can/should be used with the setting
'venv_reuse_build_only: yes'. Which will not continue to the
install stage.

Change-Id: Ib6e13ca7623942f8b0663b54528599b56b05fb19