Upgrade pip/setuptools/wheel in the venv
To ensure a uniform experience across distributions
and from build to build, we need to ensure that the
venv has a consistent version of pip, setuptools and
wheel. This patch ensures that these packages are
upgraded to the latest version allowed by the args
provided.

This prevents issues as discovered and solved by
https://review.openstack.org/596553

We also remove the virtualenv_site_packages argument
from the pip tasks as they are unnecessary given that
the virtualenv is created in an earlier task.

Change-Id: Ia9e2f695212b01a54f8e73fbd03faa8e9b492f93