Consider pre-release wheels for installation
When building from git sources or from master, pre-release
wheels may be built. We need to ensure that they are used
if they are present. For stable branches, we are still
protected by the use of the constraints to ensure that we
only get the packages we expect.

Change-Id: I68e6d29cfb6a144d119eabdeb7e23bff73c51555
(cherry picked from commit b93f331aa2ffa51f148d2d6eb2064df06501175f)