Use the same wheel path as the repo_build
Once we remove the repo build process, this needs to point
to the location where wheels are served by pypiserver.

For now we use this to ensure that we don't rebuild the wheels
that were already built by the repo build process. It has also
been found that pypiserver hangs when it encounters duplicated
wheels.

Change-Id: I86510cb7407e4ee69a376fc64ba5b8f5676f0bff