Add ability to symlink host python packages into venv
Some python packages have C bindings which tend to be very
particular about the version of their underlying shared libraries.
To ensure things run smoothly for stable releases, we opt to
use the distro packages for these python packages and symlink the
appropriate python library files and their bindings into the venv.

This functionality is required for libvirt and ceph and is used
across multiple roles.

Change-Id: Ib5b7fa1d06abe1e1bb4f14aea7de4207b61aca88