Add tripleo-ci-centos-7-standalone-os-tempest job
Since https://review.openstack.org/#/c/627500 marks the integration
of os_tempest in tripleo ci, it is good to have a job here to make
sure that it does not break tripleo integration.

python_venv_build is used in os_tempest so adding tripleo os_tempest
(non-voting) jobs makes sure that it does not break the tripleo
os_tempest integration.

Change-Id: I4d45501844c3d8fe569927fd9ccab231ab093966