Add fact gathing for hardware when building venvs
The python venv build process requires access to facts that from a
delegated host to build venvs. This change adds a fact gathering
loop to the group "repo_all" if it exists and defaults to the current
inventory hostname. This change will allow the build system to gather
all facts from all build targets as needed ensuring a successful
playbook run.

Change-Id: I8a6c60de99360570bd7ae1c1fd8055f99d770c9f
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>