Add option to upgrade PyPI/etc without using site config
There are sometimes buggy releases of PyPI shipped by the
distributions which don't react well to pip.conf config
options such extra index URLs which make it impossible to
upgrade it.

This adds a a variable which allows working around this
issue by ignoring the site config to do the upgrade.

Change-Id: I5266d827f19e14c6313b11408913e2f754befaca