speedup: move when block to outside include
Instead of including a file, then running a conditional in the block
which gets applied inside the task, which then results in every task
running, but getting skipped (especially in scenarios where the build
host is the same as the current one), we simply include if we have to.

This saves around 20 seconds in every iteration of python_venv_build
on metal.

Change-Id: I8afb2970ee696f0fe9a9e250f19f1db465d63511