Ensure the build host is the first repo server
The first repo server is the only one configured to synchronise
packages to the other repo servers. To ensure that the build
happens on the first server, we reverse the order in which they
are processed in the loop.

This does not solve the multi-architecture use-case. A solution
for that will have to follow.

Change-Id: Ie6bd16ac08164f7dba71339858b98137f61ad716