Remove the virtualenv version check
Simplify this code by making the assumption that we always have a
new enough version of virtualenv present. Centos-7 ships 15.1.0 so
it is now possible to remove some conditional logic as well.

This patch removes the version check and replaces it with a check
that the virtualenv command is present on the system patch and
is usable.

Change-Id: I1a8f4961358c9551c5493b332187b411177b2769