Replace pip requirement source specs with the package name for built wheels
When wheels are built from source, the repo server wheel build requirements
will have an input of the form git+https://example.com/foo.git@branch#egg=foo.
The wheel build will result in a wheel with a version like 1.2.3dev4.

When installing the built wheel into the venv we must switch the source
spec in the requirements from the original git repo to just the
package name 'foo', as the constraints file generated from the wheel
build will specify the exact version of the built wheel 'foo==1.2.3dev4'.

The new pip resolver will not accept there being a source spec in the
input requirements as well as a contradictory constraint from the wheel
build output, this results in an error from pip.

Change-Id: I9be6bbf4a29a4da2ddf96dc0336bc2a7d8ec9281