Add distro/arch to requirements file path
This addresses an issue when upgrading between different operating
systems. In this case, the requirements files already exist for
the old OS, which prevents new wheels from being created for the
new OS. By using a different requirements path for each OS this is
avoided.

As part of this, repeated variable construction for the
requirements path is factored our into a vars file.

Change-Id: I881a40fee31df78bf96e451509671543a49520d9