Verify that wheels are built while running against multiple hosts
Infra team has reported that OSA deployments use to DDoS gitea.
During some investigation we discovered that the potential workflow that
lead to DDoS was incorrect order during upgrade, when
compute nodes were upgraded before controllers and thus
each compute node was independently pulling same repos as no repo
server was available to this specific OS version.

Based on that we're implementing check that allows such behaviour only
if it's performed in a serial manner.
Undocumented variable `venv_wheel_build_skip_check` is added for cases
when local repos are used for service deployment and limitation should
not be applicable.

Change-Id: I54b2d4bdf6d101cc72c92be196962902a6138757