Gather build hosts facts once.
In [1] the build host facts gathering was moved from a place in the
code which was covered by run_once to one that is not. The task is
now missing a run_once so it iterates over the repo_all group for
each host in the play which could be a very large number of tasks.

[1]  I79a10a3f774246ab9c57313f19f3804f9422e08c

Change-Id: Ic86c19ca48fb55ad2c464c22f31ac41c5cc4c2e5