Drop empty elements from constraint/requirement files
We should filter out empty elements from lists that are provided as
input to the role, to ensure idempotence of build constraints and
requirements file as a result.

Additionally to that we should ensure to control jinja whitespaces
not to add extra new lines to the end of files.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible/+/872896
Change-Id: Ide47712f775c91d90b749bc8e4d8f87c5bd658f2
Closes-Bug: #2004546
(cherry picked from commit 2a25337b9844b4b8a90420e864ee38e8168ce586)