Drop unneeded become overrides
We add become overrides on the task level which neglect provided by
user options in favor of quite narrow usecase. As a result, role can not
be run as a non-root user in cases where venv_build_host is not localhost

Removing become override potenially unblocks running OSA with non-root
user on hosts.

Change-Id: I663bdb1bcc60fec0d9efc55cbb1b2a0c566ab2ab