Use uv for package installation
Introduce new variable `venv_install_tool` which allows to control
which tool (pip or uv) to use for package installation inside of the
venv.

According to tests, installation of pre-build packages which takes 5sec
with pip, takes only 0.5 sec with uv.

While it could be worth to also prepare the venv with uv, this might be
introduced with a follow up, as then uv should be
installed system-wide.

Change-Id: I4a72df1f9863b8ae6e450b46e912b82ec699333f
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>